+++ inject/inject-core/src/main/scala/com/twitter/inject/TypeUtils.scala
@@ -2,7 +2,6 @@ package com.twitter.inject
 
 import com.google.inject.TypeLiteral
 import com.google.inject.internal.MoreTypes.ParameterizedTypeImpl
-import scala.reflect.api.TypeCreator
 import scala.reflect.{ClassTag, ManifestFactory}
 import scala.reflect.runtime.universe._
 
@@ -54,37 +53,4 @@ object TypeUtils {
     }
     manifestFromType(t.tpe).asInstanceOf[Manifest[T]]
   }
-
-  /**
-   * Convert from the given Class[T] to a TypeTag[T]
-   * @param clazz the class for which to build the resultant TypeTag
-   * @return a TypeTag[T] representing the given Class[T]
-   */
-  def asTypeTag[T](clazz: Class[_ <: T]): TypeTag[T] = {
-    val clazzMirror = runtimeMirror(clazz.getClassLoader)
-    val tpe = clazzMirror.classSymbol(clazz).toType
-    val typeCreator = new TypeCreator() {
-      def apply[U <: scala.reflect.api.Universe with scala.Singleton](
-        m: scala.reflect.api.Mirror[U]
-      ): U#Type = {
-        if (clazzMirror != m) throw new RuntimeException("wrong mirror")
-        else tpe.asInstanceOf[U#Type]
-      }
-    }
-    TypeTag[T](clazzMirror, typeCreator)
-  }
-
-  /**
-   * If the given [[java.lang.reflect.Type]] is parameterized, return an Array of the
-   * type parameter names. E.g., `Map<T, U>` returns, `Array("T", "U")`.
-   */
-  private[twitter] def parameterizedTypeNames(`type`: java.lang.reflect.Type): Array[String] =
-    `type` match {
-      case pt: java.lang.reflect.ParameterizedType =>
-        pt.getActualTypeArguments.map(_.getTypeName)
-      case tv: java.lang.reflect.TypeVariable[_] =>
-        Array(tv.getTypeName)
-      case _ =>
-        Array.empty
-    }
 }